feat: add ExternalRenderer & DataPanel
