+++ src/camera/Camera.ts
@@ -152,7 +152,7 @@ abstract class Camera {
   public canReach(panel: Panel): boolean {
     const range = this._range;
 
-    if (panel.element.parentElement !== this._el) return false;
+    if (panel.removed) return false;
 
     const panelPos = panel.position;
