+++ src/core/index.ts
@@ -2,13 +2,17 @@
  * Copyright (c) 2015 NAVER Corp.
  * egjs projects are licensed under the MIT license
  */
-import Panel from "./Panel";
+import Panel, { PanelOptions } from "./Panel";
+import ElementPanel, { ElementPanelOptions } from "./ElementPanel";
 import Viewport from "./Viewport";
 import FlickingError from "./FlickingError";
 import AnchorPoint from "./AnchorPoint";
 
 export {
   Panel,
+  PanelOptions,
+  ElementPanel,
+  ElementPanelOptions,
   Viewport,
   FlickingError,
   AnchorPoint
